# docs/plan-v2/api/openapi.yaml
openapi: 3.1.0
info:
  title: HMB 온라인 — server-java API
  version: "0.1.0"
  license:
    name: Proprietary
    identifier: "UNLICENSED"
  description: |
    권위 게임 서버(Java/Spring Boot) API 계약. SoT는 이 파일이다(springdoc이 생성하는 스펙은
    구현 확인용 — 드리프트는 계약 테스트로 검출한다, LLD-server-java.md §1).

    요구사항/AC = docs/plan-v2/PRD-v2.md §3.2~3.6. DB = docs/plan-v2/ERD.md. 구현 상세 =
    docs/plan-v2/LLD-server-java.md.

    - 의존 방향: web → server-java → (ts-servants → engine). 웹은 이 API만 안다.
    - `/api/**` 는 `Authorization: Bearer <token>` 필요(로그인 자체는 제외). `/internal/**` 는
      서번트 전용 `X-Servant-Token`.
    - 에러 규약(전 엔드포인트 공통): `ApiError` 스키마. `code` 값은 `ErrorCode` 열거 참조.
    - `TacticalInput` / `SelectData` / `MatchLog` 는 `packages/shared` (zod)가 상세를 소유한다.
      여기서는 필드 설명이 있는 느슨한 오브젝트로만 정의(계약 상세는 zod가 SoT) — 필드명은
      zod 스키마와 반드시 일치시킨다(camelCase).
servers:
  - url: http://localhost:8080
    description: local dev
security:
  - bearerAuth: []
...
paths:
...
  /api/matches:
    post:
      tags: [matches]
      operationId: createMatch
      summary: 매치 생성(BRIEFING 진입) — 봇 매칭 + 활성 덱 스냅샷 + 시드 생성
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchRequest"
      responses:
        "201":
          description: 생성됨(state=BRIEFING)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "400":
          description: |
            code=DECK_INVALID(활성 덱이 유효하지 않음, AC-S2 규칙 재검증) ·
            code=DECK_REQUIRED(활성 덱 자체가 없음, #319 — 화면은 덱 구성으로 보낸다) ·
            code=TUTORIAL_UNAVAILABLE(#493 W6-v3 — `tutorial:true` 인데 배포에 구운 자산이 없다.
            **막다른 길이 아니다**: 클라는 플래그를 빼고 한 번 더 보내 일반 연습경기로 착지한다).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: |
            code=MATCH_IN_PROGRESS — 끝나지 않은 매치가 이미 있음(#217 AC2),
            detail={matchId,state,action} — 그 매치로 이어가거나 포기해야 새로 만들 수 있다.
            code=TUTORIAL_ALREADY_PLAYED — 튜토리얼 고정 매치는 1회뿐(#493 W6-v3), detail={played}.
            판정은 **FINISHED 인 튜토리얼 매치 존재**이고 ABANDONED·FAILED 는 재시도를 막지 않는다
            (구운 로그는 언제나 크게 이기므로 반복 생성이 열려 있으면 승리 보상이 무한 발행된다).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/matches/{id}:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    get:
      tags: [matches]
      operationId: getMatch
      summary: |
        매치 상태 조회(폴링용). BRIEFING이면 opponent(상대 분석) 포함. state별로 노출 필드가
        다르다 — 상세는 MatchDetail 스키마 설명 참고(§5.1).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/matches/{id}/prompts:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: submitMatchPrompt
      summary: 팀/선수 프롬프트 입력(phase=pre|halftime) — 같은 대상 재입력은 UPSERT
      description: |
        허용 상태(그 외 409 INVALID_STATE): phase=`pre` → BRIEFING /
        phase=`halftime` → **FIRST_HALF 또는 HALFTIME**(P4-E2 #170 — 전반을 보면서 후반 지시를
        미리 써두고, 감독시간에 마저 고칠 수 있다).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchPromptRequest"
      responses:
        "200":
          description: 저장됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/kickoff:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: kickoffMatch
      summary: 킥오프 — 양팀 half1 AI 잡 enqueue, BRIEFING → GEN1
      responses:
        "202":
          description: 잡 큐잉됨(state=GEN1). 클라이언트는 GET /api/matches/{id} 폴링.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/halves/{half}/log:
    parameters:
      - $ref: "#/components/parameters/MatchId"
      - name: half
        in: path
        required: true
        schema:
          type: integer
          enum: [1, 2]
    get:
      tags: [matches]
      operationId: getHalfLog
      summary: half별 MatchLog(재생용) — AC-M3, zod MatchLog 스키마와 계약
      description: |
        허용 상태(그 외 409): half=1 → FIRST_HALF·HALFTIME·GEN2·SECOND_HALF·FINISHED /
        half=2 → **SECOND_HALF·FINISHED**.

        로그는 해당 하프 **전체**를 내려준다. 라이브 재생 중 "앞서가기 금지"는 `MatchDetail.clock`
        기반 **클라 강제**다(P4-E2 R3 — 서버 절단은 PvP 백로그, 루트 §7 PvP-ready 경계와 정합).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: 매치/half 없음 또는 아직 시뮬 전(state 미도달)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/halftime:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: submitHalftimeSubs
      summary: 하프타임 교체 저장(≤3, 벤치→선발만, resume 시 반영) — §5.4, AC-M4
      description: |
        허용 상태(그 외 409): **FIRST_HALF·HALFTIME**(P4-E2 #170 — 전반 중 미리 짜두고 감독시간에 확정).
        상태 전이는 없다(후반 시뮬에서 반영).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HalftimeRequest"
      responses:
        "200":
          description: 저장됨(상태 전이는 없음 — resume에서 반영)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "400":
          description: |
            교체 규칙 위반(3명 초과·벤치 외 선수·GK 퇴장 후 GK 0). code=SUBSTITUTION_INVALID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/resume:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: resumeMatch
      summary: 후반 시작 — half2 AI 잡 enqueue(전반 요약 컨텍스트 포함), HALFTIME → GEN2
      description: |
        허용 상태 = **HALFTIME 만**. FIRST_HALF 에서 호출하면 409 — 이 엔드포인트로 후반 잡을
        앞당겨 큐잉하는 것은 계속 금지다(P4-D1).

        ⚠️ **#421 스킵과의 관계**: 재생 장면을 건너뛰고 싶으면 `POST /api/matches/{id}/skip` 이다.
        스킵은 "후반을 앞당기는 전이"가 아니라 **재생 창을 닫는 것**이고, 그 뒤는 아래에 적힌
        만료 전이가 그대로 밟는다(전이 엣지 신설 0). 즉 P4-D1 이 금지한 것 중 **재생 창을 유저가
        닫는 것**만 hero 지시로 열렸고(롤백 스위치 `hmb.match.skip.enabled`), `/resume` 의 상태
        규칙은 한 줄도 바뀌지 않았다.

        감독시간(`clock.phaseEndsAt`)이 만료되면 **서버가 같은 전이를 자동 수행**한다
        (`hmb.match.clock.auto-resume-on-expiry`). 이때 하프타임 프롬프트가 없으면 **전반 인풋이
        그대로 승계**된다(AI 콜 0 — AC-W2-2).
      responses:
        "202":
          description: 잡 큐잉됨(state=GEN2)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/skip:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: skipMatchPhase
      summary: 경기 스킵 — 재생 중인 하프의 창을 지금으로 당긴다 (#421)
      description: |
        **허용 상태** = `FIRST_HALF` · `SECOND_HALF`(= 재생 창이 열려 있는 단계). 그 외는 **409**.

        **새 상태 전이는 없다.** 서버는 `phase_ends_at` 을 `now` 로 **당기기만** 하고, 그 자리에서
        **기존 만료 전이**를 밟는다 — 전반 스킵이면 감독시간 개시(오토 #249 면 같은 체인으로 후반
        재생까지), 후반 스킵이면 종료·정산. 응답은 **전이 후** `MatchDetail`(새 `clock` 포함)이다.
        스위퍼 주기(1s)를 기다리지 않는 이유는 "닫으면 바로 다음"이 이 기능의 요구이기 때문이다.

        **바디 `phase` 는 필수이고 CAS 키다.** 서버는 `UPDATE ... WHERE state = :phase` 로만 창을
        당긴다. 전반 막바지 스킵과 스위퍼의 감독시간 개시가 1초 안에 겹칠 수 있는데, phase 가 없으면
        재전송 한 번이 **다음 단계를 통째로 날린다**. 그래서 지금 도는 단계와 다른 phase 는 409 다
        (이미 전이된 뒤의 재전송도 마찬가지 — 조용한 200 을 주지 않는다).

        **되감기 불가**: 이미 지난 창은 당기지 않는다(`AND phase_ends_at > now`). 창을 **늘리는**
        방향으로는 이 API 를 쓸 수 없다 — 지는 경기의 시간벌기 차단.

        **결정론 무영향**: 하프 로그·`resumeState`·`lastHash` 는 창이 열리기 **전에** 확정되고
        `phase_ends_at` 은 시뮬 입력 어디에도 들어가지 않는다. 스킵으로 결과가 달라질 경로는 0개다.

        **보상**: 후반 스킵의 정산은 기존 경계 CAS 를 그대로 타므로 **정확히 1회**다.

        운영 스위치 `hmb.match.skip.enabled=false` 면 **409**(창 무접촉) = P4-D1 원 규칙으로 롤백.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phase]
              properties:
                phase:
                  type: string
                  enum: [FIRST_HALF, SECOND_HALF]
                  description: 유저가 지금 보고 있는 단계. 현재 상태와 다르면 409.
      responses:
        "200":
          description: OK — 전이 후 매치 상태(clock 포함)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "400":
          description: phase 누락 또는 라이브 단계가 아닌 값. code=VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/auto:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: setMatchAuto
      summary: 오토 모드 on/off — 켜면 전반 종료 시 감독시간 없이 후반 바로 진행 (#249)
      description: |
        **허용 상태** = `BRIEFING`, `GEN1`, `FIRST_HALF`, `HALFTIME`(+레거시 `H1_BREAK`).
        후반이 이미 열린 뒤(`GEN2` 이후)는 **409** — 감독시간은 지나갔고 되돌릴 수 없다.

        플래그는 **전반 종료 경계에서만** 읽힌다. 켜져 있으면 서버가 감독시간을 **0초**로 열어
        같은 스윕 안에서 `GEN2` 로 잇는다(전이 엣지 신설 없음). 꺼져 있으면 현행
        `clock.halftimeMs` 그대로다.

        **후반 인풋에 새 AI 흐름은 없다** — 감독시간 만료와 같은 전이를 타므로 전반 진입 직후
        선행 생성된 후반 인풋(#193 W2b-B2 프리페치)을 그대로 쓴다. 아직 준비되지 않았으면
        `GEN2` 에서 대기한다(= 유저가 [후반 시작]을 즉시 눌렀을 때와 동일). 전반을 보며 미리 써 둔
        **후반 지시는 그대로 반영된다** — 오토는 지시 포기가 아니다.

        **경합**: 감독시간이 이미 열린 뒤 `auto=true` 가 도착하면 그 자리에서 후반을 연다
        (`POST /resume` 과 같은 전이). 덕분에 토글이 경계 직전/직후 어디에 떨어져도 결과가 같다.

        운영 스위치 `hmb.match.auto.enabled=false` 면 경계가 이 플래그를 보지 않는다(전부 정상
        감독시간). 그래도 이 API 는 200 을 준다 — 롤백이 클라 에러로 새지 않게.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [auto]
              properties:
                auto: { type: boolean }
      responses:
        "200":
          description: OK — 반영된 매치 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/result:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    get:
      tags: [matches]
      operationId: getMatchResult
      summary: 최종 결과(스코어·팀/선수 스탯·리포트). FINISHED에서만 의미 있음(읽기 전용) — §5.5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"
  /api/matches/{id}/abandon:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: abandonMatch
      summary: |
        매치 포기(#217 AC3) — ACTIVE → ABANDONED. 잠금(유저당 미완 매치 1개)의 탈출구다:
        이게 없으면 고아 매치 하나가 계정을 영구히 잠근다.

        허용 조건은 서버가 판정한다(리롤 방지): BRIEFING(킥오프 전) · FAILED(생성 실패) ·
        시계가 멈춘 라이브(phase_ends_at + stuck-grace 경과). **정상 재생 중에는 409** —
        지고 있는 경기를 버리고 다시 뽑는 리롤을 막는다. 지금 포기 가능한지는
        GET /api/me/active-match 의 abandonable 로 알 수 있다.
      responses:
        "200":
          description: 회수됨(state=ABANDONED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 포기할 수 없는 상태(정상 재생 중이거나 이미 종료). code=INVALID_STATE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/matches/{id}/retry:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    post:
      tags: [matches]
      operationId: retryMatch
      summary: FAILED 매치의 실패 지점 재큐잉 — AC-M7
      responses:
        "202":
          description: 재큐잉됨(실패 직전 GEN* 상태로 복귀)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidState"

  # ───────────────────────────── internal (서번트 전용) ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "/api/** 전체(로그인 제외). sessions 테이블 조회."
    servantToken:
      type: apiKey
      in: header
      name: X-Servant-Token
      description: "/internal/** 전용 고정 shared secret(AC-Q3)."
  parameters:
    MatchId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: 인증 실패(토큰 없음/만료/무효). code=UNAUTHORIZED
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: 리소스 없음. code=NOT_FOUND
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ValidationError:
      description: 입력 검증 실패. code=VALIDATION_ERROR
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    InvalidState:
      description: |
        현재 매치 상태에서 허용되지 않는 액션(§5.1 전이표 외). code=INVALID_STATE. AC-M1.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: 사람이 읽을 수 있는 사유(한국어 가능)
        detail:
          type: [object, "null"]
          additionalProperties: true
      example:
        code: DECK_INVALID
        message: 선발이 11명이 아닙니다
        detail:
          starterCount: 10

    # ───────────────────────────── auth/me ─────────────────────────────
...
    MatchDetail:
      type: object
      description: |
        state에 따라 일부 필드만 채워진다: BRIEFING → opponent 포함. HALFTIME(구 H1_BREAK) 이후 →
        scoreH1Home/Away. FINISHED → scoreHome/Away+result. FAILED → failReason.
        라이브 단계(FIRST_HALF/HALFTIME/SECOND_HALF) → clock(P4-E2 #170).
        후반 스코어는 **FINISHED 전까지 노출하지 않는다**(재생 중 스포일러 금지).
      required: [id, state, createdAt]
      properties:
        id: { type: string }
        state:
          $ref: "#/components/schemas/MatchState"
        failReason:
          type: [string, "null"]
        opponent:
          $ref: "#/components/schemas/OpponentAnalysis"
        scoreH1Home: { type: [integer, "null"] }
        scoreH1Away: { type: [integer, "null"] }
        scoreHome: { type: [integer, "null"] }
        scoreAway: { type: [integer, "null"] }
        result:
          type: [string, "null"]
          enum: [WIN, DRAW, LOSS]
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
        clock:
          description: |
            서버 권위 시계(P4-E2 #170, additive). 라이브 단계가 아니거나 시계 비활성이면 null.
          oneOf:
            - $ref: "#/components/schemas/MatchClock"
            - type: "null"
        auto:
          type: boolean
          description: |
            오토 모드(#249, additive). true 면 전반 종료 시 감독시간 없이 후반이 바로 시작된다.
            변경은 `POST /api/matches/{id}/auto`. 구 매치·기본값은 false.
        tutorial:
          type: boolean
          description: |
            온레일 튜토리얼 고정 매치인가(#493 W6-v3, additive). 구 매치·기본값은 false.
            web 은 이 값 하나로 **재생 정지·[스킵] 잠금·하프타임 교체 차단**을 건다 — 후반 로그가
            미리 구워져 있어 교체·자리 지시가 결과에 반영되지 않기 때문이다(반영 안 되는 조작을
            열어 두면 유저는 자기 지시가 사라졌다고 읽는다).
            ⚠️ 관전자 경로(`toDetailFor`)에서는 `auto` 와 같은 이유로 **상수 false**.

            ⚠️ 이 값이 true 여도 운영 스위치(`hmb.match.auto.enabled=false`)가 내려가 있으면 흐름은
            정상 감독시간이다 — 그건 롤백 상태이고 클라가 알 필요는 없다.
    CreateMatchRequest:
      type: object
      properties:
        botId:
          type: [string, "null"]
          description: "생략 시 랜덤 봇(BOT_ATK/BOT_DEF/BOT_BAL 중, PRD §3.4)"
        tutorial:
          type: [boolean, "null"]
          description: |
            온레일 튜토리얼 고정 매치로 만든다(#493 W6-v3, additive). 생략·false = 기존 연습경기.
            true 면 서버가 **미리 구운 로그**를 적재한다 — AI 호출 0 · 러너 왕복 0 · 대기 0 ·
            전 유저 bit-identical 한 유저 승리(고정 시드). ⚠️ `botId` 는 무시된다(상대는 자산이 정한다).
            `mode` 는 여전히 `practice` 이고 표식만 다르다(모드 값을 늘리면 보상 곡선·전적·리그
            분기가 "모르는 모드"로 폴백한다).
            거절 2종은 **실패가 아니라 일반 연습경기로 가라는 뜻**이다 —
            `400 TUTORIAL_UNAVAILABLE`(배포에 자산이 없다) · `409 TUTORIAL_ALREADY_PLAYED`(1회 제한).
    MatchPromptRequest:
      type: object
      required: [phase, scope, text]
      properties:
        phase:
          type: string
          enum: [pre, halftime]
        scope:
          type: string
          enum: [team, player]
        playerId:
          type: [string, "null"]
          description: "scope='player'일 때 필수"
        text:
          type: string
          maxLength: 500
    SubstitutionItem:
      type: object
      required: [out, "in"]
      properties:
        out:
          type: string
          description: playerId(전반 선발)
        "in":
          type: string
          description: playerId(벤치)
    HalftimeRequest:
      type: object
      required: [substitutions]
      properties:
        substitutions:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/SubstitutionItem"
          description: |
            ≤3(hmb.match.halftime-subs-max), out∈전반 선발, in∈벤치, 교체 후 GK≥1.
            **생략 = 손대지 않음**(앞서 낸 교체 유지) / **빈 배열 = 교체 전부 취소**(명시적 의사).
      description: |
        ⚠️ Phase2 추가 필드 — teamTactics(#254 — 감독시간 팀 전술) · formation + starters
        (#276 — 감독시간 포메이션·선발 배치) — 는 openapi-v2 `HalftimeRequestPhase2Fields` 가
        소유한다(스펙 2개는 각각 독립 생성물이라 교차 $ref 금지).
    MatchResult:
      type: object
      required: [matchId, scoreHome, scoreAway, result]
      properties:
        matchId: { type: string }
        scoreHome: { type: integer }
        scoreAway: { type: integer }
        result:
          type: string
          enum: [WIN, DRAW, LOSS]
        pointsAwarded:
          type: integer
          description: 승/무/패 보상(economy.v1.json 수치, 멱등 지급 — AC-M6)
        teamStats:
          type: object
          additionalProperties: true
          description: 팀 스탯 요약(슛/점유율 등) — 상세는 W3 구현에서 확정
        playerStats:
          type: array
          items:
            type: object
            additionalProperties: true
          description: 선수별 스탯 리스트 — 상세는 W3 구현에서 확정

    # ───────────────────────────── shared 계약(느슨한 stub) ─────────────────────────────
...
    MatchLog:
      type: object
      description: |
        packages/shared/src/match-log.ts (zod SoT). 결정론 엔진 출력 — 틱별 좌표 스냅샷 +
        이벤트 시계열 + 최종 스코어 + 틱 해시(재현/desync 검증). 필드: configVersion, seed,
        tickSnapshots(TickSnapshot[]: tick·minute·ball·ballOwner·players·hash),
        events(MatchEvent[]: tick·minute·type·team?·playerId?·xg?·detail?), finalScore
        {home,away}. AC-M3: 이 응답은 zod MatchLog로 파싱 가능해야 한다.
      required: [configVersion, seed, tickSnapshots, events, finalScore]
      properties:
        configVersion:
          type: string
        seed:
          type: string
        tickSnapshots:
          type: array
          items:
            type: object
            additionalProperties: true
        events:
          type: array
          items:
            type: object
            additionalProperties: true
        finalScore:
          type: object
          properties:
            home: { type: integer }
            away: { type: integer }

    # ───────────────────────────── internal ai-jobs ─────────────────────────────
...